requirements.txt
openai==1.16.0
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv